Consolidate project files
This consolidates the multiple project files we have right now, embellishing
further the Lumi.yaml/json project file as the place for all things project
and package related.  Although I suspect we will revist some of this during
next sprint's package management work, this is a step in the right direction.

As a result, we can rip out all the tsconfig.json files scattered throughout.

The new structure adds two top-level sections to the Lumi.yaml/json file:

    language:
        compiler: lumixx
        settings:
            a: foo
            b: bar
            z: baz
    files:
        - index.xx
        - config.xx

The idea here is that the manifest self-describes three things: 1) the compiler
used to build the package, 2) the settings to pass to that compiler (an opaque
bag to Lumi, since these are specific to the compiler), and 3) the source files
that comprise the project.  This helps to unify the different compilers.

This accomplishes work item pulumi/lumi#49.  It also enables pulumi/lumi#98,
the `lumi compile` and automatic recompilation commands, which will come soon.