Support captured reference to `this` in lambdas
Lumi returns references to the `this` variable as
`.this`, so we need to rewrite these when emitting
into Lambdas.  In addition, the TypeScript compilation
of lambda functions requires that we create an environment
for executing the generated code where `this` has the
right value, which requires some subtle changes to our code
generation.